refactoring FooterSocial to SetSocial in BtnReserved
