+++ src/components/Footer/BtnReserved/BtnReserved.module.scss
@@ -1,0 +1,32 @@
+@import '@css/index.scss';
+
+.btnBoxReserved {
+  & {
+    display: flex;
+    justify-content: center;
+  }
+
+  @include tablet {
+    display: none;
+  }
+}
+
+.btnReserved {
+  & {
+    border: none;
+    background-color: $bg-btn;
+    border-radius: 15px;
+    padding: 15px 38px;
+    margin: 20px 0;
+  }
+
+  @include desktop {
+    margin: 28px 0 20px 0;
+  }
+}
+
+.iconFooterColor {
+  min-width: 42px;
+  min-height: 42px;
+  color: $icon-social-footer;
+}
