+++ src/components/Footer/Footer.module.scss
@@ -61,3 +61,9 @@
     padding: 0;
   }
 }
+
+.iconFooterColor {
+  min-width: 42px;
+  min-height: 42px;
+  color: $icon-social-footer;
+}
