+++ src/components/Footer/FooterAddress/FooterAddress.module.scss
@@ -2,13 +2,17 @@
 @import '@font/fonts.scss';
 
 .addressBox {
-  @include tablet {
-    justify-content: start;
-    padding: 0;
-  }
-
   & {
     padding: 32px 0 0 0;
+  }
+
+  @include tablet {
+    justify-content: start;
+    padding: 32px 0;
+  }
+
+  @include desktop {
+    padding: 0;
   }
 }
