+++ src/components/Footer/FooterContact/FooterContact.module.scss
@@ -68,7 +68,7 @@
   @include tabletMin {
     @include font(16px, 300, 00.01em);
     & {
-      margin: 32px 0 0 0;
+      margin: 20px 0 0 0;
     }
 
     &:active {
