--- src/components/Footer/FooterSocial/FooterSocial.module.scss
@@ -1,79 +1,0 @@
-@import '@css/index.scss';
-
-// .socialFooterContainer {
-    
-//     @include desktop{
-//         display: flex;
-//         justify-content: end;
-//         align-items: center;
-//     }
-  
-// }
-
-.btnBoxReserved {
-  & {
-    display: flex;
-    justify-content: center;
-  }
-
-  @include tablet {
-    display: none;
-  }
-
-//   @include desktop {
-//     display: block;
-//   }
-}
-
-.btnReserved {
-  & {
-    border: none;
-    background-color: $bg-btn;
-    border-radius: 15px;
-    padding: 15px 38px;
-    margin: 20px 0;
-  }
-
-  @include tabletMin {
-    @include font(16px, 300, 00.01em);
-    & {
-      margin-bottom: 0;
-    }
-
-    &:active {
-      box-shadow: inset 0 0 2px 0px rgba(0, 0, 0, 0.5);
-    }
-  }
-}
-
-.socialFooter {
-  & {
-    display: flex;
-    justify-content: center;
-    padding-bottom: 32px;
-  }
-
-  @include tabletMin {
-    justify-content: end;
-    // padding: 0;
-    padding-top: 32px;
-    margin: 0;
-  }
-}
-
-.socialFooterItem {
-  width: 42px;
-  height: 42px;
-  margin-right: 32px;
-  padding: 0;
-
-  &:last-child {
-    margin-right: 0;
-  }
-}
-
-.footerIconColor {
-  min-width: 42px;
-  min-height: 42px;
-  color: $icon-social-footer;
-}
