--- src/components/Footer/FooterSocial/index.js
@@ -1,1 +1,0 @@
-export { default } from './FooterSocial';
